requirements.txt
uv
python-dotenv
kagglehub
pandas
numpy
langchain-kuzu
langchain-community
langchain-google-genai
uv